Almost compiles

+++ firmware/HelperClasses/HSVtoRGB.h
@@ -4,9 +4,9 @@
 
 typedef struct
 {
-    int R;
-    int G;
-    int B;
+    uint8_t R;
+    uint8_t G;
+    uint8_t B;
 }RGB_t;
 
 RGB_t HSVtoRGB(uint8_t hue, uint8_t sat, uint8_t val);
