+++ firmware/HelperClasses/WindowAvg.c
@@ -5,7 +5,8 @@
 /* Private function declarations */
 
 /* Public function definitions */
-void WndAvg_init(WndAvg_t* pObj, int32_t* buf, uint16_t len)
+
+void WndAvg_init(WndAvg_t* pObj, int16_t* buf, uint8_t len)
 {
     pObj->buf = buf;
     pObj->len = (len > 0) ? len : 1;    //guard against len == 0
@@ -16,12 +17,12 @@ void WndAvg_clear(WndAvg_t* pObj)
 {
     pObj->sum = 0;
     pObj->idx = 0;
-    for (uint16_t i = 0; i < pObj->len; i++) {
+    for (uint8_t i = 0; i < pObj->len; i++) {
         pObj->buf[i] = 0;
     }
 }
 
-void WndAvg_pushVal(WndAvg_t* pObj, int32_t val)
+void WndAvg_pushVal(WndAvg_t* pObj, int16_t val)
 {
     pObj->sum += val - pObj->buf[pObj->idx];    // Add new val, subtract old val
     pObj->buf[pObj->idx] = val;                 // Replace old val
